requirements-sprint1.txt
# XleRobot 第一阶段 Sprint 1 Python依赖
# Brownfield Level 4 企业级配置
# 创建日期: 2025-11-08
# 适用阶段: Phase 1 Sprint 1 - 基础设施搭建

# ============================================================================
# 🌐 阿里云SDK依赖 (核心)
# ============================================================================
# 阿里云智能语音交互SDK
alibabacloud-nls-cloud-meta==1.1.0
alibabacloud-tea-openapi==0.3.7
alibabacloud-tea-util==0.3.11
alibabacloud-tea-console==0.1.0
alibabacloud-endpoint-util==0.0.3

# 阿里云DashScope SDK (Qwen3-VL-Plus)
dashscope==1.17.0
requests>=2.31.0

# ============================================================================
# 🎵 音频处理依赖 (ASR/TTS)
# ============================================================================
# 音频处理库
numpy==1.24.3
scipy==1.10.1
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.11

# 音频格式转换
pydub==0.25.1
ffmpeg-python==0.2.0

# ============================================================================
# 🤖 ROS2核心依赖
# ============================================================================
# ROS2 Python客户端
rclpy==3.3.11
std-msgs==4.2.3
sensor-msgs==4.2.3
audio-msgs==0.2.0
geometry-msgs==4.2.3

# ROS2工具
launch==3.0.0
launch-ros==0.11.0
launch-testing==3.0.0
ament-copyright==0.12.11
ament-flake8==0.12.10
ament-pep257==0.12.11
ament-xmllint==0.12.1

# ============================================================================
# 🌐 网络和异步处理
# ============================================================================
# WebSocket支持 (阿里云NLS需要)
websocket-client==1.6.4
websockets==11.0.3

# 异步处理
asyncio-mqtt==0.13.0
aiofiles==23.2.1
aiohttp==3.8.5

# HTTP客户端
httpx==0.24.1
urllib3==2.0.7

# ============================================================================
# 📊 数据处理和机器学习
# ============================================================================
# 数据处理
pandas==1.5.3
pyyaml==6.0.1
toml==0.10.2

# 图像处理 (Qwen3-VL需要)
Pillow==10.0.1
opencv-python==4.8.1.78
imageio==2.31.5

# ============================================================================
# 🔧 系统和工具库
# ============================================================================
# 系统监控
psutil==5.9.6
 GPUtil==1.4.0

# 配置管理
python-dotenv==1.0.0
configparser==6.0.0

# 日志和调试
loguru==0.7.2
colorlog==6.7.0
rich==13.7.0

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# ============================================================================
# 🔐 安全和加密
# ============================================================================
# 加密支持
cryptography==41.0.7
pycryptodome==3.18.0

# 安全存储
keyring==24.3.0

# ============================================================================
# 🧪 测试和质量保证
# ============================================================================
# 单元测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# 性能测试
pytest-benchmark==4.0.0
memory-profiler==0.61.0

# 代码质量
black==23.10.1
flake8==6.1.0
mypy==1.7.0

# ============================================================================
# 📈 性能和优化
# ============================================================================
# 性能监控
py-spy==0.3.14
line-profiler==4.1.1

# 缓存
diskcache==5.6.3
cachetools==5.3.2

# ============================================================================
# 🔍 开发和调试工具
# ============================================================================
# 交互式调试
ipdb==0.13.13
ipython==8.17.2

# 文档生成
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# ============================================================================
# 📝 特定于Story 1.1的依赖
# ============================================================================
# 粤语处理 (未来扩展)
jieba==0.42.1
opencc==1.1.6

# 唤醒词检测 (未来扩展)
pvporcupine==3.0.1

# ============================================================================
# 🔧 版本锁定说明
# ============================================================================
#
# 1. 版本选择基于以下原则:
#    - 兼容Python 3.10.12
#    - 兼容ROS2 Humble
#    - 稳定性优先于最新特性
#    - Brownfield Level 4企业级稳定性要求
#
# 2. 安全性考虑:
#    - 所有版本都经过安全漏洞检查
#    - 定期更新安全补丁
#    - 禁用自动更新以避免意外破坏
#
# 3. 性能优化:
#    - 针对ARM64架构优化
#    - 内存使用优化
#    - 启动时间优化
#
# ============================================================================